+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next
+- Improve text selection in modal
+- Change spacing units from `rem` to `px`
+
 ## 1.0.1 - 2022 Dec 12
 - Fix usage instructions